🚧 Explore converting json to xml
When converting JSON to XML, the process works much smoother, the _root_
is set and the hierarchy works though I _think_ the magic sauce is in
the child_component_count_isi field.  I'm going to try again with
getting JSON to work and adding that to see if the hierarchy will work.