⚙️ Make sure assets reload in development
This commit will add the listen gem and configure the development.rb's
file watcher to make for faster assets reloading.  To make that happen
we gotta use `bin/dev` to start the server, but I'm using to using
`bin/web` so I'm just calling `bin/dev` from `bin/web`.  By default, it
runs the server on http://127.0.0.1:3000 but my localhost is on
http://0.0.0.0:3000 so I edited the Procfile.dev to use that instead.

If assets don't reload, you can try `rails assets:clobber` and restart
the server.