🎁 Add faceted search for evolutions fields
Now when we go to the card show page, if it has an "Evolves To" or
"Evolves From" field, it will be a link to a faceted search of that
Pokemon.  I changed the evolves from field to ssm just to make it easy
and match the evovles to field.