🎁 Sort by card number
This will make it so we are sorting by the card number.  I needed to
change the sort from sort_isi to sort_ssi and using the card number as
the value.  The default sort_isi wasn't working because we can't 100% go
off of the order of the JSON response so we want to sort it by the card
number instead.  However, it needs to be a string because some card
numbers have string characters.  Padding the string with 8 zeros should
do the trick, we'll see.