⚙️ Configure Blacklight Gallery
This will add a new thumbnail_path_ssi index field for both sets and
cards.  For the sets, this will be the logo and the cards will be the
small version of the cards.  I don't think we need the masonry and
slideshow views from Blacklight Gallery so I commented that out.  Also,
the compact view was rather pointless as well.