🎁 Add new component to display the image on show
This will rework the way the metadata displays on the set and card show
pages by using a different section for the image and the rest of the
indexed metadata.  This allows us to use a separate component to display
the image.  Right now it's just a div with the image inside.

Also, I overrode the `Arclight::MetadataSectionComponent` so we don't
get the headers and it looks more like the card show page.