🐛 Fix logic for pokemon search
When we do an all fields search it now returns not just the card with
the Pokemon name because sometimes attack/ability text would reference
the name of the Pokemon.  I tweaked it so that it now searches for the
Pokemon name.  Also slowly adding back search options, starting with the
"Card Name".