🎁 Improve UI with storing y position
The scroll_position_controller is completely AI and after testing it,
it's actually working pretty good.  This should help prevent the UI from
jumping around too much when you're clicking on the facets, which was
getting annoying.  I override the catalog index page so I can add the
data-controller attribute to.