💄 Use bg-primary in navbar
Since the navbar uses bg-dark by default, I'm just overriding the color
to use bg-primary instead.  This is probably the easiest way as opposed
to actually overriding the `Blacklight::TopNavbarComponent` to remove
the bootstrap class.  Also, I added a logo-dark since the bg-primary is
actually darker than the bg-dark so the text would stand out against the
darker background.