🧹 Adjust bottom border in grouped view
In the grouped view, we had a bottom border that was only hidden for the
last item but when we were in a wider view where there are three items
per row, this looked odd.  This CSS should fix that when we are in a
larger width, we hide the last three items but only if we're in the
gallery view.