🎁 Use a custom card list on card show page sidebar
Prior, we were using the default Arclight collection context list which
lists out all the components of the collection.  However, our data isn't
going to be nearly as big and is not nested.  This way we don't need to
wrap it in a turbo frame and it will load much faster.