🎁 Turn on bookmarks and make it easily configured
This commit will add a config where we can set bookmarks to true and it
would work without going into too many places.  Also had to comment out
the `OnlineStatusIndicatorComponent` because we don't need it and it
would actually take up some space in the title area.  We can also get
rid of the custom `DocumentComponent` now because we don't need it
anymore.