🐛 Fix infinite scroll losing position
Previously, when going to the show page and then hitting the back
button, the user would lose their spot on the page if we did any
infinite scrolling on the index page.  This was because with the POST
request in the default title a tag, something like:

`<a data-context-href="/catalog/zsv10pt5-97/track?counter=97&amp;document_id=zsv10pt5-97&amp;per_page=30&amp;search_id=643" data-context-method="post" data-turbo-prefetch="false" itemprop="name" href="/catalog/zsv10pt5-97">Darumaka</a>`

it would make the back button from the show page do a server request,
which means if we had 90 documents loaded with infinite scroll, it would
now be reduced back down to the default 30.  If the user was on the 75th
document, for example, then it wouldn't return them back to that spot
since we only have 30 results.

This commit will remove the POST and just use the direct link to the
show page and thus not creating the odd behavior.