🧹 Compensate for API errors
This is definitely not an idea way but I figure since it's a brand new
set I can temporarily correct any errors I seen in the API.  The Victini
is coming as 'Rare' instead of 'Black White Rare' from the API and there
was a number problem on Antique Cover Fossil as well.  Hopefully, this
gets corrected in the API and this can be reverted.

Ref:
- https://api.pokemontcg.io/v2/cards/zsv10pt5-171
- https://api.pokemontcg.io/v2/cards/rsv10pt5-172
- https://api.pokemontcg.io/v2/cards/zsv10pt5-80