🎁 Add infinite scroll to Bookmarks index
This commit will pull the log for infinite scroll out to a concern so we
can use it in the Bookmarks index.  I had to create an alias because it
wouldn't find the method otherwise.  The index template was brought over
so we can add the stimulus controllers onto it.