🎁 Add IIIF Viewer link to Bookmark
This commit will add the IIIF Viewer link to the bookmark page to view
the bookmarked items in Mirador.  This will also allow users to share
their bookmarked items in the viewer.  I decided to use the hashid gem
to encode the user id so it's not just a sequential number.  It's not
really a secret or anything but just makes it less prone to guessing.