🎁 Add ActionCable support for bookmarks
This commit looks big but a lot of files were generated by Rails using:
```sh
rails generate channel bookmarks
```

I did delete the bookmarks_channel file it generated because I put the
websocket logic in the sortable_controller instead.  So what this
effectively does is makes it so if a user has two (or more) tabs of
their bookmarks index open, it will update the bookmarks across all tabs
otherwise the order may become unintended if the user loses track of
which tab is has the more updated order.  Mainly, this is so I can play
with ActionCable since the use case is pretty niche, but not unlikely.