🐛 Add fix for when there isn't a logo or symbol
The TCGDex API is missing some data in their returns so this is just a
quick fix.  We can assume the logo and symbol image urls if they are not
present in the return data.  It's still not perfect because if the image
actually doesn't exist then it won't load.  Might be time to ask TCGDex
for a fix.